+++ tasmota/language/zh_TW.h
@@ -243,6 +243,7 @@
 
 // xdrv_02_webserver.ino
 #define D_NOSCRIPT "為了要使用 Tasmota，請啟用 JavaScript"
+#define D_SAFEBOOT "SAFEBOOT"
 #define D_MINIMAL_FIRMWARE_PLEASE_UPGRADE "MINIMAL韌體<br>請升級"
 #define D_WEBSERVER_ACTIVE_ON "網頁伺服器已經啟動，位於:"
 #define D_WITH_IP_ADDRESS "IP位址:"
@@ -372,6 +373,7 @@
 #define D_OTA_URL "OTA網址"
 #define D_START_UPGRADE "開始升級"
 #define D_UPGRADE_BY_FILE_UPLOAD "透過檔案升級"
+#define D_UPLOAD_FACTORY "Switching to safeboot partition"
 #define D_UPLOAD_STARTED "已開始上傳"
 #define D_UPGRADE_STARTED "已開始升級"
 #define D_UPLOAD_DONE "已上傳完成"
@@ -865,7 +867,8 @@
 #define D_GPIO_SHIFT595_OE     "74x595 OE"
 #define D_GPIO_SHIFT595_SER    "74x595 SER"
 #define D_SENSOR_CM11_TX       "CM110x TX" 
-#define D_SENSOR_CM11_RX       "CM110x RX" 
+#define D_SENSOR_CM11_RX       "CM110x RX"
+#define D_SENSOR_FLOWRATEMETER "Flowrate"
 
 // Units
 #define D_UNIT_AMPERE "安培"
@@ -913,6 +916,8 @@
 #define D_UNIT_WATT "瓦特"
 #define D_UNIT_WATTHOUR "瓦小時"
 #define D_UNIT_WATT_METER_QUADRAT "W/m²"
+#define D_UNIT_LITER_PER_MINUTE "l/min"
+#define D_UNIT_CUBICMETER_PER_HOUR "m³/h"
 
 #define D_NEW_ADDRESS          "Setting address to"
 #define D_OUT_OF_RANGE         "Out of Range"
@@ -1053,6 +1058,9 @@
 #define D_FP_ADDRCODE "Address code"           // 0x20 Address code
 #define D_FP_PASSVERIFY "Password verified"    // 0x21 Verify the fingerprint passed
 #define D_FP_UNKNOWNERROR "Error"              // Any other error
+
+// xsns_96_flowratemeter.ino
+#define D_FLOWRATEMETER_NAME "Flowrate"
 
 // xsns_83_neopool.ino
 #define D_NEOPOOL_MACH_NONE               "NeoPool"           // Machine names