+++ CHANGELOG.md
@@ -6,15 +6,30 @@ All notable changes to this project will be documented in this file.
 ## [11.0.0] 20220212
 - Release Olivia
 
-## [11.1.0.1]
+## [11.1.0.2]
+### Added
+- ESP32 Command ``Restart 3`` to switch between SafeBoot and Production 
+
+### Changed
+- ESP32 Tasmota SafeBoot with changed partition scheme allowing larger binaries
+
+### Fixed
+
+
+### Removed
+
+
+## [11.1.0.1] 20220504
 ### Added
 - Support for Sonoff MS01 soil moisture sensor (#15335)
 - Support for daisy chaining MAX7219 displays (#15345)
+- Support for light schemes via DDP as default for ESP32x (#15436)
 - Command ``EnergyExportActive<phase>`` to (p)reset energy export active for supported devices. Currently ADE7880 only (#13515)
 - Sonoff SPM delayed SetPowerOnState (#13447)
 - Command ``SetOption139 0/1`` to switch between pressure unit "mmHg" (0) or "inHg" (1) when ``SO24 1`` (#15350)
-
-### Changed
+- Support for flowrate meters like YF-DN50 and similary (#15474)
+- Command ``IfxRp ""|<policy>`` adds optional InfluxDb Retention Policy (#15513)
+- Command ``SetOption140 0/1`` to switch between MQTT Clean Session (0) or Persistent Session (1) (#15530)
 
 ### Fixed
 - SCD40 start low power command (#15361)
@@ -22,7 +37,7 @@ All notable changes to this project will be documented in this file.
 - BL09xx negative power presentation (#15374)
 
 ### Removed
-
+- Arduino IDE support
 
 ## [Released]